Add support for new vehicle color system and mark old ushort id methods as obsolete

+++ tools/UnturnedImages.Module/UI/UIManager.cs
@@ -244,7 +244,7 @@ namespace UnturnedImages.Module.UI
             }
         }
 
-        private void OnExportModClicked(ISleekElement button, uint modId)
+        private void OnExportModClicked(ISleekElement button, ulong modId)
         {
             var vehicleAngles = new Vector3(
                 _vehicleAnglesXInput?.Value ?? 0,
